Functional trash refactor

+++ TaskPanelLibrary/Service/PanelService.cs
@@ -9,13 +9,13 @@ using Task = TaskPanelLibrary.Entity.Task;
 namespace TaskPanelLibrary.Service;
 
 public class PanelService : IPanelService
-{ 
+{
     private readonly IPanelRepository _panelRepository;
 
     private IUserService _userService;
-    
+
     private ITrashService _trashService;
-    
+
     public PanelService(IPanelRepository panelRepository, IUserService userService, ITrashService trashService)
     {
         _panelRepository = panelRepository;
@@ -27,7 +27,7 @@ public class PanelService : IPanelService
     {
         if (!IsValidPanel(panel))
             throw new PanelNotValidException("Panel is not valid");
-        
+
         var panelSaved = _panelRepository.AddPanel(panel);
         return panelSaved;
     }
@@ -38,11 +38,10 @@ public class PanelService : IPanelService
         {
             return _panelRepository.GetAllPanels().FindAll(i => i.Team.Id == idTeam);
         }
-        catch(ArgumentException e)
+        catch (ArgumentException e)
         {
             return new List<Panel>();
         }
-
     }
 
     public List<Panel> GetAllPanelForUser(int userId)
@@ -50,9 +49,9 @@ public class PanelService : IPanelService
         try
         {
             User user = _userService.GetUserById(userId);
-            return _panelRepository.GetAllPanels().FindAll(i =>  i.Team.Users.Contains(user)).ToList();
+            return _panelRepository.GetAllPanels().FindAll(i => i.Team.Users.Contains(user)).ToList();
         }
-        catch(ArgumentException e)
+        catch (ArgumentException e)
         {
             return new List<Panel>();
         }
@@ -69,42 +68,49 @@ public class PanelService : IPanelService
     public Panel DeletePanel(int panelId, User user)
     {
         var panel = _panelRepository.GetPanelById(panelId);
-        
+
         if (!user.IsAdmin)
         {
             throw new PanelNotValidException($"User is not admin, userId: {user.Id}");
         }
-        
-        _panelRepository.DeletePanel(panelId);
-        _trashService.AddPanelToTrash(panel, user.TrashId);
+
+        panel.IsDeleted = true;
+
+        if (!_trashService.IsFull(user.TrashId)){
+            _trashService.AddPanelToTrash(panel, user.TrashId);
+            _panelRepository.UpdatePanel(panel);
+        }
+        else
+            _panelRepository.DeletePanel(panelId);
         
         return panel;
     }
+
 
     public Panel GetPanelById(int panelId)
     {
         return _panelRepository.GetPanelById(panelId);
     }
-    
+
     public List<Panel> GetAllPanels()
     {
         return _panelRepository.GetAllPanels();
     }
-    
+
     private bool IsValidPanel(Panel panel)
     {
         if (panel == null)
             throw new PanelNotValidException("Panel is null");
-        
+
         if (string.IsNullOrEmpty(panel.Name))
             throw new PanelNotValidException("Panel name is null or empty");
-        
+
         if (string.IsNullOrEmpty(panel.Description))
             throw new PanelNotValidException("Panel description is null or empty");
-        
+
         if (panel.Team == null)
             throw new PanelNotValidException("Panel team is null");
-        
+
         return true;
     }
 }
